Change to size_t

+++ src/p2.cpp
@@ -109,8 +109,8 @@ void node_union(tree_node_t* x, tree_node_t* y) {
     link(find_set(x), find_set(y));
 }
 
-unsigned long long int get_maximum_cost_spanning_tree(heap_t* heap) {
-    unsigned long long int result = 0;
+size_t get_maximum_cost_spanning_tree(heap_t* heap) {
+    size_t result = 0;
 
     while (heap->size > 0) {   
         edge_t* edge = extract_max(heap);
